virtio: add virtio spec
This spec includes a draft device spec for the virtio-vhost-user device.
It was based on the initial device spec written be Stefan Hajnoczi. This
spec uses the initial device spec and it incorporates a few fixes and
enhancements. The initial device spec can be found here:

https://stefanha.github.io/virtio/vhost-user-slave.html#x1-2830007

The changes I made can be found here:

https://github.com/ndragazis/virtio/commits/virtio-vhost-user